Add deprecation notice to readme
There are still users on this version of the wordpress plugin and
even though there is a newer version of the plugin, not everyone knows
that and ends up trying to use this one instead. To make it easy on
people I have just added a notice to the readme as well as a link to the
newer version.

Fixes #138